+++ src/libssc-sensor-light.c
@@ -87,7 +87,7 @@ report_received (SSCClient *self, guint32 msg_id, guint64 uid_high, guint64 uid_
 	SSCSensorLight *sensor = SSC_SENSOR_LIGHT (user_data);
 	guint64 sensor_uid_low;
 	guint64 sensor_uid_high;
-	gfloat intensity;
+	gdouble intensity;
 
 	g_object_get (sensor,
 		      SSC_SENSOR_UID_HIGH, &sensor_uid_high,
